Consolidate hidden into layout package
