+++ docs/components/spark-components.ts
@@ -10,8 +10,8 @@ export { Divider } from '@spark-web/divider';
 export { Dropzone } from '@spark-web/dropzone';
 export { Field } from '@spark-web/field';
 export { Heading } from '@spark-web/heading';
-export { Hidden } from '@spark-web/hidden';
 export * from '@spark-web/icon';
+export { Hidden } from '@spark-web/layout';
 export { Container } from '@spark-web/layout';
 export { Inline } from '@spark-web/layout';
 export { Columns } from '@spark-web/layout';
