+++ docs/package.json
@@ -29,7 +29,6 @@
     "@spark-web/dropzone": "0.0.0",
     "@spark-web/field": "0.0.0",
     "@spark-web/heading": "0.0.0",
-    "@spark-web/hidden": "0.0.0",
     "@spark-web/icon": "0.0.0",
     "@spark-web/layout": "0.0.0",
     "@spark-web/link": "0.0.0",
