--- packages/hidden/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@spark-web/hidden",
-  "version": "0.0.0",
-  "repository": "https://github.com/brighte-labs/spark-web/tree/main/packages/hidden",
-  "license": "MIT",
-  "main": "dist/spark-web-hidden.cjs.js",
-  "module": "dist/spark-web-hidden.esm.js",
-  "dependencies": {
-    "@babel/runtime": "^7.14.6",
-    "@emotion/css": "^11.7.1",
-    "@spark-web/core": "0.0.0",
-    "@spark-web/heading": "0.0.0",
-    "@spark-web/text": "0.0.0",
-    "react": "^17.0.2"
-  },
-  "devDependencies": {
-    "@types/react": "^17.0.12"
-  },
-  "engines": {
-    "node": ">= 14.13"
-  }
-}
