--- packages/hidden/src/index.ts
@@ -1,5 +1,0 @@
-export { Hidden } from './Hidden';
-
-// types
-
-export type { HiddenProps } from './Hidden';
