+++ packages/layout/package.json
@@ -10,6 +10,8 @@
     "@emotion/css": "^11.7.1",
     "@spark-web/core": "0.0.0",
     "@spark-web/divider": "0.0.0",
+    "@spark-web/heading": "0.0.0",
+    "@spark-web/text": "0.0.0",
     "react": "^17.0.2"
   },
   "devDependencies": {
