+++ packages/layout/src/hidden/index.ts
@@ -1,0 +1,5 @@
+export { Hidden } from './hidden';
+
+// types
+
+export type { HiddenProps } from './hidden';
