+++ packages/layout/src/index.ts
@@ -1,5 +1,6 @@
 export { Columns } from './columns';
 export { Container } from './container';
+export { Hidden } from './hidden';
 export { Inline } from './inline';
 export { Row } from './row';
 export { Stack } from './stack';
@@ -8,6 +9,7 @@ export { Stack } from './stack';
 
 export type { ColumnsProps } from './columns';
 export type { ContainerProps } from './container';
+export type { HiddenProps } from './hidden';
 export type { InlineProps } from './inline';
 export type { RowProps } from './row';
 export type { StackProps } from './stack';
